requirements.txt
streamlit
boto3
psycopg2-binary
python-dotenv
